fix(lint): resolve pre-existing eslint errors blocking CI on 1.1.0
Unrelated to the sidebar-style PR — these were already failing on
1.1.0's frontend CI job before that PR existed:

- CategorizeTab.test.tsx: unused `_enabled` mock parameter.
- CategorizeTab.tsx / RevolutTab.tsx / AddAccountModal.tsx: silence
  react-hooks/set-state-in-effect on running->done transition
  detectors — these react to a ref-tracked transition (and, for
  RevolutTab/AddAccountModal, trigger further mutations), not derive
  render state, so the effect is the right tool here.
- TransactionDetailSheet.tsx: genuinely was deriving state from
  props in an effect — the one case this lint rule is right about.
  Rewritten to adjust state during render (tracking the (tx,
  categories) pair last derived from), matching the original
  effect's dependency semantics exactly, and dropped the now-unused
  useEffect import.
- PeriodNavigator.tsx: moved its exported pure date helpers into a
  new period-navigator-helpers.ts module to fix
  react-refresh/only-export-components, same pattern already used
  for sidebar-nav-items.ts.
